+++ CHANGELOG.md
@@ -1,7 +1,13 @@
+## 1.0.2
+
+- **FIX**: Fixed the issue with displaying the pin/unpin button on smartphones.
+- **FIX**: Resolved the toggle functionality for nested items.
+- **CHORE**: Added standard comments across the package for better code documentation.
+
 ## 1.0.1
 
-- - **FEAT**: Add web support
+- **FEAT**: Add web support.
 
 ## 1.0.0
 
-- - **FEAT**: Add initial version
+- **FEAT**: Add initial version.